Move McpConfigProvider to correct package
